fix(shadow): address #41 review — public soft-shadow API, continuous presentation parity, angular-radius scheduling
Review fixes on top of the #41 area-light implementation:

1. Public API: DomLightState.angularRadius (radians, dimensionless) and
   DomShadowOptions.samples ({1,4,8,16}) flow through buildScene,
   scaleShadowOptions (samples forwarded UNSCALED — counts never dpr-scale),
   UkiboriDom state/setLight and the React <Ukibori> props (angularRadius +
   shadow.samples), with CPU/WebGPU parity tests.

2. The #41 sample benchmark now runs the REAL soft path (benchmark scene
   carries angularRadius 0.15) and reports requested AND effective sample
   counts per row plus whether the soft path was active.

3. Continuous presentation parity: the binary vis<0.5 branch is gone from
   the CPU oracle — compositeShadowPremultipliedStrengthBytes mirrors the
   WGSL op-for-op (alpha and premultiplied RGB both scaled by
   clamp(1-vis,0,1)); hard {0,1} bytes unchanged. New real-WebGPU
   presentation fixture present-soft-shadow-custom-tint-alpha exercises
   intermediate visibility on the canvas with a custom tint/alpha at 8
   samples.

4. Scheduling: LIGHT_ANGULAR_RADIUS_REGION (header offset 88) is no longer
   part of HEADER_GEOMETRY_REGIONS (only the reserved word 92..96 stays);
   an angularRadius-only change classifies as the new light-angular-radius
   reason -> upload/shadow/lighting/presentation with height/normal
   retained; heightInputsMatchScene authorizes retained height fields;
   planFrame treats it as a semantic full-dispatch change; explicit
   executed/skipped scheduler test added.

5. Production CPU cone directions are now derived from the CANONICAL f32
   light direction (the same encoded header components the GPU reads), not
   raw f64; unit test pins dirs == header-derived dirs.

6. sanitizeAngularRadius is the single source of truth: createScene now
   rejects values whose f32 packing overflows to Infinity (NaN/+-Inf/
   negative/f32-overflow all -> 0); createScene->encode->validate roundtrip
   test added.

7. Coverage: shadow-soft-mask-caster fixture (mask/glyph caster, exact
   fractional GPU parity); catalog visibility policy description updated to
   distinguish exact-binary hard from exact-dyadic-fractional soft.
   golden-core --update now also persists intentional @metadata (policy
   table) edits; goldens regenerated accordingly.

+++ packages/renderer/src/shadow.ts
@@ -230,15 +230,27 @@ export function prepareShadowContext(
   const sampleHeight = options.casterHeight ?? height;
   // #41 soft-shadow sampling state: the light's angular radius (radians,
   // f32-packed, 0 = hard) and the sanitized sample count. The cone
-  // directions are computed ONCE here in f64/f32 exactly like the array the
-  // GPU uniform packs, so both backends march identical f32 directions.
+  // directions are derived from the CANONICAL F32 light direction — the GPU
+  // marches the encoded header's fround'ed components, so the production
+  // CPU path must round-trip through the same packing before building the
+  // basis — and computed once here exactly like the array the GPU uniform
+  // packs, so both backends march identical f32 directions. The HARD march
+  // keeps consuming the historical (unrounded) context values.
   const angularRadius = sanitizeAngularRadius(
     typeof scene.light.angularRadius === "number" ? scene.light.angularRadius : undefined,
   );
   const samples = sanitizeShadowSamples(options.samples);
   const sampleDirs =
     angularRadius > 0 && samples > 1
-      ? computeSoftSampleDirections({ x: lx, y: ly, z: lz }, angularRadius, samples)
+      ? computeSoftSampleDirections(
+          {
+            x: Math.fround(lx),
+            y: Math.fround(ly),
+            z: Math.fround(lz),
+          },
+          angularRadius,
+          samples,
+        )
       : null;
   return {
     stepSize,
